Refactor Helpers
